requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
polyline>=2.0.0
gpxpy>=1.6.0
